Move Error into server_axum module
